refactor: input map hook, move useContractQueries to hooks index
