+++ frontend/src/components/Simulator/ContractInfo.vue
@@ -1,11 +1,10 @@
 <script setup lang="ts">
-import { useContractQueries } from '@/hooks/useContractQueries';
 import PageSection from '@/components/Simulator/PageSection.vue';
 import { CheckCircleIcon } from '@heroicons/vue/24/outline';
 import EmptyListPlaceholder from '@/components/Simulator/EmptyListPlaceholder.vue';
 import { PlusIcon } from '@heroicons/vue/16/solid';
 import { useNodeStore } from '@/stores';
-import { useShortAddress } from '@/hooks';
+import { useShortAddress, useContractQueries } from '@/hooks';
 
 const nodeStore = useNodeStore();
 const { shortenAddress } = useShortAddress();
