+++ frontend/src/hooks/index.ts
@@ -5,3 +5,5 @@ export * from './useEth';
 export * from './useShortAddress';
 export * from './useDb';
 export * from './useWebSocketClient';
+export * from './useInputMap';
+export * from './useContractQueries';
