+++ frontend/src/utils/index.ts
@@ -1,3 +1,2 @@
-export * from './inputs';
 export * from './editor';
 export * from './store';
